Refactor tick_size in Instrument to use size_t for improved precision and consistency

+++ server_app/instrument/instrument.h
@@ -32,7 +32,7 @@ public:
     ShareString symbol;
     ShareString exchange_symbol;
     size_t lot_size;
-    double tick_size;
+    size_t tick_size;
 
     Json to_json() const;
     static Instrument from_json(Json& data);
